pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "saia-ingest"
description = "GeneXus Enterprise AI Ingest utilities"
authors = ["Luciano Silveira <lsilveira@genexus.com>"]
license = "MIT"
#packages = [{include = "saia-ingest"}]
readme = "README.md"
repository = "https://www.github.com/genexuslabs/saia-ingest"
version = "0.0.1"

[tool.poetry.scripts]
saia-cli = 'saia_ingest.command_line.command_line:main'


[tool.poetry.dependencies]
python = "^3.9,<3.12"
asttokens = "^2.4.1"
atlassian-python-api = "^3.41.4"
blobfile = "^2.0.1"
boto3 = "^1.34.0"
botocore = "^1.34.0"
bs4 = "^0.0.1"
cachetools = "^5.3.1"
colorama = "^0.4.6"
comm = "^0.2.0"
debugpy = "^1.8.0"
decorator = "^5.1.1"
exceptiongroup = "^1.2.0"
executing = "^2.0.1"
google-api-core = "^2.11.0"
google-auth = "^2.19.1"
google-cloud-aiplatform = "^1.26.0"
google-cloud-bigquery = "^3.11.0"
google-cloud-core = "^2.3.2"
google-cloud-resource-manager = "^1.10.1"
google-cloud-storage = "^2.9.0"
google-crc32c = "^1.5.0"
google-resumable-media = "^2.5.0"
google-search-results = "^2.4.2"
googleapis-common-protos = "^1.59.0"
greenlet = "^2.0.2"
httpcore = "^0.16.3"
httptools = "^0.5.0"
httpx = "^0.23.3"
ipykernel = "^6.27.1"
ipython = "^8.18.1"
itsdangerous = "^2.0.1"
jira = "^3.5.2"
jupyter_client = "^7.4.8"
jupyter_core = "^5.1.3"
langchain = "^0.0.350"
llama-hub = "^0.0.58"
llama-index = "^0.9.15"
matplotlib-inline = "^0.1.6"
openai = "^1.3.9"
openapi-schema-pydantic = "^1.2.4"
packaging = "^23.2"
pinecone-client = "^2.2.4"
pipreqs = "^0.4.11"
psutil = "^5.9.6"
pure-eval = "^0.2.2"
Pygments = "^2.17.2"
pytest = "^7.4.3"
python-dateutil = "^2.8.2"
python-dotenv = "^0.19.2"
pytz = "^2022.7.1"
pywin32 = "^306"
PyYAML = "^6.0"
pyzmq = "^25.1.2"
regex = "^2022.10.31"
tiktoken = "^0.5.2"
tokenizers = "^0.15.0"
tqdm = "^4.66.1"
